requirement.txt
fastapi
uvicorn[standard]
httpx
python-dotenv
pydantic